Updated install script - with temporary version

+++ install-checker-suite.sh
@@ -5,10 +5,21 @@
 # Usage:
 # ------
 #
-# ./install-checker-suite.sh <base_directory>
+# ./install-checker-suite.sh [--no-conda] [<base_directory>]
 #
 
-CHECKS_BASE_DIR=${1:-${PWD}/checks-work-dir}
+INSTALL_DIR=
+INSTALL_CONDA=1
+
+while [[ "$#" -gt 0 ]]; do
+    case $1 in
+        -n|--no-conda) INSTALL_CONDA=0 ;;
+        *) INSTALL_DIR=$1 ;;
+    esac
+    shift
+done
+
+CHECKS_BASE_DIR=${INSTALL_DIR:-${PWD}/checks-work-dir}
 CHECKS_BASE_DIR=$(realpath $CHECKS_BASE_DIR)
 
 ENV_NAME=amf-checks-env
@@ -22,12 +33,23 @@ fi
 echo "[INFO] Making/checking base directory: $CHECKS_BASE_DIR"
 mkdir -p $CHECKS_BASE_DIR
 
-echo "[INFO] Installing miniconda and python..."
-cd $CHECKS_BASE_DIR/
-wget https://repo.anaconda.com/miniconda/Miniconda3-latest-Linux-x86_64.sh
-sh Miniconda3-latest-Linux-x86_64.sh -p ${CHECKS_BASE_DIR}/miniconda3 -b
-rm -f Miniconda3-latest-Linux-x86_64.sh
-export PATH=$PATH:${CHECKS_BASE_DIR}/miniconda3/bin
+if [ $INSTALL_CONDA -eq 0 ]; then
+    which conda > /dev/null 2>&1
+    if [ $? -ne 0 ]; then
+        echo "[ERROR] Cannot find 'conda' executable. Suggested fix: run with '--no-conda' option."
+        exit
+    fi 
+    echo "[INFO] Found 'conda' executable, will use the local installation."
+
+else
+    echo "[INFO] Installing miniconda and python..."
+    cd $CHECKS_BASE_DIR/
+
+    wget https://repo.anaconda.com/miniconda/Miniconda3-latest-Linux-x86_64.sh
+    sh Miniconda3-latest-Linux-x86_64.sh -p ${CHECKS_BASE_DIR}/miniconda3 -b
+    rm -f Miniconda3-latest-Linux-x86_64.sh
+    export PATH=$PATH:${CHECKS_BASE_DIR}/miniconda3/bin
+fi
 
 echo "[INFO] Creating environment: $ENV_NAME"
 conda create --name $ENV_NAME python=3.9 -y
@@ -36,34 +58,25 @@ source activate $ENV_NAME
 echo "[INFO] Install the third-party compliance-checker framework..."
 conda install -c conda-forge compliance-checker pip -y
 
+CHECKS_VERSION=v2.0.1test
+MINOR_VERSION=$(echo $CHECKS_VERSION | cut -d. -f1-2)
+
 echo "[INFO] Install and/or clone the relevant repositories..."
-pip install git+https://github.com/cedadev/compliance-check-lib
-pip install git+https://github.com/cedadev/cc-yaml
-pip install git+https://github.com/ncasuk/amf-check-writer
+CEDADEV_REPOS="cedadev/compliance-check-lib cedadev/cc-yaml"
+NCASUK_REPOS="ncasuk/amf-check-writer ncasuk/AMF_CVs.git@${CHECKS_VERSION} ncasuk/amf-compliance-checks.git@${CHECKS_VERSION}"
 
-CV_VERSIONS=2.0.0
-PACKAGES="AMF_CVs amf-compliance-checks"
-
-for pkg in $PACKAGES; do
-    zfile=v${CV_VERSIONS}.zip
-    wget https://github.com/ncasuk/${pkg}/archive/refs/tags/${zfile}
-    unzip $zfile
-    rm -f $zfile 
+for repo in $CEDADEV_REPOS $NCASUK_REPOS; do
+    pip install git+https://github.com/$repo
 done
 
 echo "[INFO] Create setup file..."
-PYESSV_ARCHIVE_HOME=$CHECKS_BASE_DIR/AMF_CVs-${CV_VERSIONS}/amf-pyessv-vocabs
-CHECKS_DIR=$CHECKS_BASE_DIR/amf-compliance-checks-${CV_VERSIONS}/amf-checks
-
 setup_file=${CHECKS_BASE_DIR}/setup-checks-env.sh
 
 echo "export CHECKS_BASE_DIR=$CHECKS_BASE_DIR" >> $setup_file
 echo "export PATH=\$PATH:\${CHECKS_BASE_DIR}/miniconda3/bin" >> $setup_file
 echo " "  >> $setup_file
 echo "source activate amf-checks-env" >> $setup_file
-echo "export PYESSV_ARCHIVE_HOME=$PYESSV_ARCHIVE_HOME" >> $setup_file
-echo "export CHECKS_DIR=$CHECKS_DIR" >> $setup_file
-echo "export VERSION=v2.0" >> $setup_file
+echo "export CHECKS_VERSION=${MINOR_VERSION}" >> $setup_file
 
 echo "[INFO] To setup environment, do:"
 echo "source $setup_file" 
@@ -73,10 +86,10 @@ echo "[INFO] You can test it with:"
 
 echo "source $setup_file"
 echo "TEST_FILE_NAME=ncas-anemometer-1_ral_29001225_mean-winds_v0.1.nc"
-echo 'TEST_FILE_URL="https://github.com/cedadev/compliance-check-lib/blob/master/tests/example_data/nc_file_checks_data/${TEST_FILE_NAME}?raw=true"'
+echo 'TEST_FILE_URL="https://github.com/cedadev/compliance-check-lib/blob/main/tests/example_data/nc_file_checks_data/${TEST_FILE_NAME}?raw=true"'
 echo "wget -O \$TEST_FILE_NAME \$TEST_FILE_URL"
-echo "amf-checker --yaml-dir \$CHECKS_DIR --version \$VERSION \$TEST_FILE_NAME"
+echo "amf-checker --version \$CHECKS_VERSION \$TEST_FILE_NAME"
 
 echo "[INFO] Or more generally:"
-echo "amf-checker --yaml-dir $CHECKS_DIR --version <version> <test_file>"
+echo "amf-checker --version <version> <test_file>"
